refactor(describegpt): extract dictionary formatters into submodule
Move format_dictionary_markdown / format_dictionary_json /
format_dictionary_tsv and their shared helper extract_ordered_addl_cols
into src/cmd/describegpt/formatters.rs. These four functions had no
dependency on describegpt's runtime state — they are pure transforms
from &[DictionaryEntry] to String/Value — so they split cleanly.

Also decouple format_dictionary_json from the Args struct by taking the
three flags it cared about (enum_threshold, num_examples, truncate_str)
as primitives, so the submodule has no import-back into the parent
beyond DictionaryEntry itself.

No behavior change. 234 lines moved out of describegpt.rs; clippy clean,
all 64 integration + 7 unit tests pass.
